fix: circular dependency, introduce rollup, cleanup GH actions
