+++ package.json
@@ -23,18 +23,27 @@
   },
   "license": "MIT",
   "author": "Nathan Hopkins <natehop@gmail.com>",
-  "main": "./javascript/index.js",
-  "module": "./javascript/index.js",
+  "main": "./dist/cable_ready.umd.js",
+  "module": "./dist/cable_ready.module.js",
+  "files": [
+    "dist/*",
+    "src/*"
+  ],
   "scripts": {
     "lint": "yarn run prettier-standard:check",
     "format": "yarn run prettier-standard:format",
     "prettier-standard:check": "yarn run prettier-standard --check ./javascript/**/*.js",
-    "prettier-standard:format": "yarn run prettier-standard ./javascript/**/*.js"
+    "prettier-standard:format": "yarn run prettier-standard ./javascript/**/*.js",
+    "build": "yarn rollup -c",
+    "watch": "yarn rollup -wc"
   },
   "dependencies": {
     "morphdom": "^2.6.1"
   },
   "devDependencies": {
-    "prettier-standard": "^16.4.1"
+    "@rollup/plugin-commonjs": "^21.0.1",
+    "@rollup/plugin-node-resolve": "^13.0.6",
+    "prettier-standard": "^16.4.1",
+    "rollup": "^2.60.0"
   }
 }
